+++ src/main/java/com/ripple/BE/post/application/CommentCommandUseCase.java
@@ -1,4 +1,4 @@
-package com.ripple.BE.post.service;
+package com.ripple.BE.post.application;
 
 public interface CommentCommandUseCase {
