+++ src/main/java/com/ripple/BE/post/application/CommentLikeUseCase.java
@@ -1,4 +1,4 @@
-package com.ripple.BE.post.service;
+package com.ripple.BE.post.application;
 
 public interface CommentLikeUseCase {
     void addLikeToComment(final long commentId, final long userId, final long postId);
