+++ src/main/java/com/ripple/BE/post/application/PostCommandUseCase.java
@@ -1,7 +1,7 @@
-package com.ripple.BE.post.service;
+package com.ripple.BE.post.application;
 
-import com.ripple.BE.post.service.command.CreatePostCommand;
-import com.ripple.BE.post.service.command.UpdatePostCommand;
+import com.ripple.BE.post.application.command.CreatePostCommand;
+import com.ripple.BE.post.application.command.UpdatePostCommand;
 
 public interface PostCommandUseCase {
