+++ src/main/java/com/ripple/BE/post/application/ToktokAdminUseCase.java
@@ -1,4 +1,4 @@
-package com.ripple.BE.post.service;
+package com.ripple.BE.post.application;
 
 public interface ToktokAdminUseCase {
     void createToktokByExcel();
