+++ src/main/java/com/ripple/BE/post/application/impl/common/PostLikeService.java
@@ -1,14 +1,14 @@
-package com.ripple.BE.post.service.impl.common;
+package com.ripple.BE.post.application.impl.common;
 
 import static com.ripple.BE.post.exception.errorcode.PostErrorCode.*;
 
 import com.ripple.BE.notification.service.NotificationService;
+import com.ripple.BE.post.application.PostLikeUseCase;
 import com.ripple.BE.post.domain.post.Post;
 import com.ripple.BE.post.domain.post.PostLike;
 import com.ripple.BE.post.exception.PostException;
 import com.ripple.BE.post.persistence.PostLikeRepository;
 import com.ripple.BE.post.persistence.PostRepository;
-import com.ripple.BE.post.service.PostLikeUseCase;
 import com.ripple.BE.user.domain.User;
 import com.ripple.BE.user.service.UserService;
 import lombok.RequiredArgsConstructor;
