+++ src/main/java/com/ripple/BE/post/application/impl/common/PostScrapService.java
@@ -1,13 +1,13 @@
-package com.ripple.BE.post.service.impl.common;
+package com.ripple.BE.post.application.impl.common;
 
 import static com.ripple.BE.post.exception.errorcode.PostErrorCode.*;
 
+import com.ripple.BE.post.application.PostScrapUseCase;
 import com.ripple.BE.post.domain.post.Post;
 import com.ripple.BE.post.domain.post.PostScrap;
 import com.ripple.BE.post.exception.PostException;
 import com.ripple.BE.post.persistence.PostRepository;
 import com.ripple.BE.post.persistence.PostScrapRepository;
-import com.ripple.BE.post.service.PostScrapUseCase;
 import com.ripple.BE.user.domain.User;
 import com.ripple.BE.user.service.UserService;
 import lombok.RequiredArgsConstructor;
