+++ src/main/java/com/ripple/BE/post/controller/CommentController.java
@@ -1,8 +1,8 @@
 package com.ripple.BE.post.controller;
 
 import com.ripple.BE.global.dto.response.ApiResponse;
+import com.ripple.BE.post.application.CommentCommandUseCase;
 import com.ripple.BE.post.dto.request.CommentRequest;
-import com.ripple.BE.post.service.CommentCommandUseCase;
 import com.ripple.BE.user.domain.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
