+++ src/main/java/com/ripple/BE/search/controller/SearchController.java
@@ -3,10 +3,10 @@ package com.ripple.BE.search.controller;
 import com.ripple.BE.global.dto.response.ApiResponse;
 import com.ripple.BE.news.dto.NewsListDTO;
 import com.ripple.BE.news.dto.response.NewsListResponse;
+import com.ripple.BE.post.application.PostQueryUseCase;
+import com.ripple.BE.post.application.ToktokQueryUseCase;
 import com.ripple.BE.post.dto.response.PostPreviewListResponseDTO;
 import com.ripple.BE.post.dto.response.ToktokPreviewListResponseDTO;
-import com.ripple.BE.post.service.PostQueryUseCase;
-import com.ripple.BE.post.service.ToktokQueryUseCase;
 import com.ripple.BE.search.dto.SearchKeywordListDTO;
 import com.ripple.BE.search.dto.response.SearchKeywordListResponse;
 import com.ripple.BE.search.service.SearchService;
